Fix: Pet meters no longer go over 100%. - Split ratelimit stuff from Utils and create RatelimitUtils - Some small fixes.
