+++ src/main/java/net/kodehawa/mantarobot/commands/OwnerCmd.java
@@ -57,12 +57,12 @@ import net.kodehawa.mantarobot.utils.data.JsonDataManager;
 
 import java.awt.*;
 import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
+import java.util.Date;
 import java.util.List;
-import java.util.*;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
-
-import static net.kodehawa.mantarobot.utils.StringUtils.SPLIT_PATTERN;
 
 @Module
 public class OwnerCmd {
